Refactor: Move command modules to commands/ subdirectory
Moved all <name>_commands.py modules from ethcli/ to ethcli/commands/<name>.py.
Updated imports in ethcli/cli.py accordingly.
