+++ ethcli/commands/test.py
